Move DatabaseHelpers to the dnp3mocks library
