+++ cpp/tests/dnp3mocks/include/dnp3mocks/DatabaseHelpers.h
