+++ cpp/tests/unit/TestOutstation.cpp
@@ -18,9 +18,9 @@
  * limitations under the License.
  */
 #include "utils/APDUHexBuilders.h"
-#include "utils/DatabaseHelpers.h"
 #include "utils/OutstationTestObject.h"
 
+#include <dnp3mocks/DatabaseHelpers.h>
 #include <ser4cpp/util/HexConversions.h>
 
 #include <catch.hpp>
